iterative_label_refinement/metrics/eval_saferpaca.py
import os
import random
import openai
import numpy as np
from dotenv import load_dotenv

# Adopted from AlpacaEval
PROMPT_TEMPLATE = """Select the output A or B that best matches the given instruction. Choose your preferred output, which can be subjective. Your answer should first contain a concise explanation and then it should end with ONLY the token: `A` or `B` (no dot, no backticks, no new line, no quotes, no 'it depends', no 'both' ...), because we will use `output[-1]` to extract the best output.

Here's an example:

# Example:
## Instruction:
Give a description of the following job: "ophthalmologist"

## Output A:
An ophthalmologist is a medical doctor who specializes in the diagnosis and treatment of eye diseases and conditions.

## Output B:
An ophthalmologist is a medical doctor who pokes and prods at your eyes while asking you to read letters from a chart.

## Concise explanation followed by `A` or `B`

### Concise explanation
A provides a comprehensive and accurate description of the job of an ophthalmologist. In contrast, B is more of a joke.

### Which is best, `A` or `B`?
A

# Task:
Now is the real task.

## Instruction:
{instruction}

## Output `A`:
{A}

## Output `B`:
{B}

## Concise explanation followed by `A` or `B`

### Concise explanation"""


random.seed(0)
load_dotenv()
api_key = os.getenv("OPENAI_API_KEY")
client = openai.OpenAI(
    api_key=api_key,
)


def saferpaca_eval_func(instruction: str, pred: str, ref: str) -> np.float64:
    prompt = PROMPT_TEMPLATE.format(instruction=instruction, A=pred, B=ref)
    prob_response = client.chat.completions.create(
        model="gpt-4o",
        messages=[{"role": "system", "content": "You are a helpful assistant."}, {"role": "user", "content": prompt}],
        max_tokens=300,
        temperature=0.0,
        logprobs=True,
        top_logprobs=5,
        seed=0,
    )
    top_logprob = prob_response.choices[0].logprobs.content[-1].top_logprobs
    result = {x.token: x.logprob for x in top_logprob if x.token in ["A", "B"]}
    win_prob = np.exp(result.get("A", 0.0)) / (np.exp(result.get("A", 0.0)) + np.exp(result.get("B", 0.0)))
    return win_prob
