requirements.txt
accelerate==0.31.0
datasets==2.20.0
fire==0.6.0
func_timeout==4.3.5
huggingface-hub==0.23.4
ipython==8.12.3
numpy==1.26.4
openai==1.35.3
pandas==2.2.2
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sqlparse==0.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.2
vllm==0.5.0.post1
wandb==0.17.2
peft==0.11.1
python-dotenv==1.0.1
